jsarcase/test/conformance2/rendering/framebuffer-unsupported.html

<!--
Copyright (c) 2019 The Khronos Group Inc.
Use of this source code is governed by an MIT-style license that can be
found in the LICENSE.txt file.
-->

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>WebGL FRAMEBUFFER_UNSUPPORTED Test</title>
<link rel="stylesheet" href="../../resources/js-test-style.css"/>
<script src="../../js/js-test-pre.js"></script>
<script src="../../js/webgl-test-utils.js"></script>
<script src="../../js/webgl-async-loader.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>
<canvas id="canvas" width="2" height="2"> </canvas>

<script type="module">
"use strict";
import * as loader from '../../js/webgl-async-loader.js';
const { waitForWebGLTestGlobals } = loader;

// å°†åŸå§‹æµ‹è¯•ä»£ç å°è£…ä¸ºå‡½æ•°
function runWebGLTest() {
  try {

var wtu = WebGLTestUtils;
var gl;
var canvas = document.getElementById("canvas");

function checkFramebuffer(expected) {
    var actual = gl.checkFramebufferStatus(gl.FRAMEBUFFER);
    if (expected.indexOf(actual) < 0) {
        var msg = "checkFramebufferStatus expects [";
        for (var index = 0; index < expected.length; ++index) {
            msg += wtu.glEnumToString(gl, expected[index]);
            if (index + 1 < expected.length)
                msg += ", ";
        }
        msg += "], was " + wtu.glEnumToString(gl, actual);
        testFailed(msg);
    } else {
        var msg = "checkFramebufferStatus got " + wtu.glEnumToString(gl, actual) +
                  " as expected";
        testPassed(msg);
    }
}

function testImageAttachedTwoPoints() {
    debug("");
    debug("Checking an image is attached to more than one color attachment in a framebuffer.");

    var tex1 = gl.createTexture();
    var tex2 = gl.createTexture();
    var fb = gl.createFramebuffer();
    gl.bindTexture(gl.TEXTURE_2D, tex1);
    gl.texImage2D(gl.TEXTURE_2D,
                  0,                                          // level
                  gl.RGBA,                                    // internalFormat
                  1,                                          // width
                  1,                                          // height
                  0,                                          // border
                  gl.RGBA,                                    // format
                  gl.UNSIGNED_BYTE,                           // type
                  new Uint8Array(4));                         // data
    gl.bindTexture(gl.TEXTURE_2D, tex2);
    gl.texImage2D(gl.TEXTURE_2D, 0, gl.RGBA, 1, 1, 0, gl.RGBA, gl.UNSIGNED_BYTE, new Uint8Array(4));
    wtu.glErrorShouldBe(gl, gl.NO_ERROR, "Texture creation should succeed.");

    gl.bindFramebuffer(gl.FRAMEBUFFER, fb);
    gl.framebufferTexture2D(gl.FRAMEBUFFER, gl.COLOR_ATTACHMENT0, gl.TEXTURE_2D, tex1, 0);
    checkFramebuffer([gl.FRAMEBUFFER_COMPLETE]);
    gl.framebufferTexture2D(gl.FRAMEBUFFER, gl.COLOR_ATTACHMENT1, gl.TEXTURE_2D, tex2, 0);
    checkFramebuffer([gl.FRAMEBUFFER_COMPLETE]);
    gl.framebufferTexture2D(gl.FRAMEBUFFER, gl.COLOR_ATTACHMENT2, gl.TEXTURE_2D, tex1, 0);
    checkFramebuffer([gl.FRAMEBUFFER_UNSUPPORTED]);

    gl.bindFramebuffer(gl.FRAMEBUFFER, null);
    gl.deleteFramebuffer(fb);
    fb = gl.createFramebuffer();
    gl.bindFramebuffer(gl.FRAMEBUFFER, fb);
    var texCube = gl.createTexture();
    gl.bindTexture(gl.TEXTURE_CUBE_MAP, texCube);
    for (var target = gl.TEXTURE_CUBE_MAP_POSITIVE_X; target < gl.TEXTURE_CUBE_MAP_POSITIVE_X + 6; target++) {
      gl.texImage2D(target, 0, gl.RGBA, 1, 1, 0, gl.RGBA, gl.UNSIGNED_BYTE, new Uint8Array(4));
    }
    gl.framebufferTexture2D(gl.FRAMEBUFFER, gl.COLOR_ATTACHMENT0, gl.TEXTURE_CUBE_MAP_POSITIVE_X, texCube, 0);
    checkFramebuffer([gl.FRAMEBUFFER_COMPLETE]);
    gl.framebufferTexture2D(gl.FRAMEBUFFER, gl.COLOR_ATTACHMENT1, gl.TEXTURE_CUBE_MAP_POSITIVE_Y, texCube, 0);
    checkFramebuffer([gl.FRAMEBUFFER_COMPLETE]);
    gl.framebufferTexture2D(gl.FRAMEBUFFER, gl.COLOR_ATTACHMENT2, gl.TEXTURE_CUBE_MAP_POSITIVE_X, texCube, 0);
    checkFramebuffer([gl.FRAMEBUFFER_UNSUPPORTED]);


    gl.bindFramebuffer(gl.FRAMEBUFFER, null);
    gl.deleteFramebuffer(fb);
    fb = gl.createFramebuffer();
    gl.bindFramebuffer(gl.FRAMEBUFFER, fb);
    var tex3d = gl.createTexture();
    gl.bindTexture(gl.TEXTURE_3D, tex3d);
    gl.texImage3D(gl.TEXTURE_3D,
                  0,                                          // level
                  gl.RGBA,                                    // internalFormat
                  2,                                          // width
                  2,                                          // height
                  2,                                          // depth
                  0,                                          // border
                  gl.RGBA,                                    // format
                  gl.UNSIGNED_BYTE,                           // type
                  new Uint8Array(4 * 2 * 2 * 2));             // data
    gl.framebufferTextureLayer(gl.FRAMEBUFFER, gl.COLOR_ATTACHMENT0, tex3d, 0, 0);
    checkFramebuffer([gl.FRAMEBUFFER_COMPLETE]);
    gl.framebufferTextureLayer(gl.FRAMEBUFFER, gl.COLOR_ATTACHMENT1, tex3d, 0, 1);
    checkFramebuffer([gl.FRAMEBUFFER_COMPLETE]);
    gl.framebufferTextureLayer(gl.FRAMEBUFFER, gl.COLOR_ATTACHMENT2, tex3d, 0, 0);
    checkFramebuffer([gl.FRAMEBUFFER_UNSUPPORTED]);

    // Clean up
    gl.deleteTexture(tex1);
    gl.deleteTexture(tex2);
    gl.deleteTexture(texCube);
    gl.deleteTexture(tex3d);
    gl.deleteFramebuffer(fb);
}

description("This tests FRAMEBUFFER_UNSUPPORTED.");

shouldBeNonNull("gl = wtu.create3DContext(undefined, undefined, 2)");

testImageAttachedTwoPoints();

debug("");

    var successfullyParsed = true;
    console.log("âœ… WebGL test completed successfully!");
  } catch (error) {
    console.error("âŒ WebGL test failed:", error);
    if (typeof debug !== 'undefined') {
      debug("Test failed: " + error.message);
    }
  }
}

// ç¯å¢ƒæ£€æµ‹å‡½æ•°
function isJSARRuntime() {
  const checks = [
    () => typeof waitForWebGLTestGlobals !== 'undefined',
    () => typeof process !== 'undefined' && process.versions && process.versions.node,
    () => typeof navigator !== 'undefined' && navigator.userAgent && navigator.userAgent.includes('JSAR'),
    () => typeof global !== 'undefined' && global.process,
    () => typeof window === 'undefined' && typeof global !== 'undefined',
    () => typeof module !== 'undefined' && module.exports
  ];
  
  const result = checks.some(check => {
    try {
      return check();
    } catch (e) {
      return false;
    }
  });
  
  console.log('ğŸ” JSAR Runtime detection result:', result);
  return result;
}

// ç®€å•è½®è¯¢ç­‰å¾…å‡½æ•°
function waitForGlobalsSimple(globals, callback, maxAttempts = 300) {
  let attempts = 0;
  console.log('ğŸ”„ Starting simple polling for globals:', globals);
  
  function check() {
    attempts++;
    const available = [];
    const missing = [];
    
    globals.forEach(name => {
      try {
        if (typeof eval(name) !== 'undefined') {
          available.push(name);
        } else {
          missing.push(name);
        }
      } catch (e) {
        missing.push(name);
      }
    });
    
    if (missing.length === 0) {
      console.log('âœ… All globals available:', globals);
      callback();
    } else if (attempts >= maxAttempts) {
      console.error('âŒ Timeout waiting for globals. Available:', available, 'Missing:', missing);
    } else {
      if (attempts % 20 === 0 || attempts <= 5) {
        console.log(`â³ Waiting for globals... Available: [${available.join(', ')}], Missing: [${missing.join(', ')}] (${attempts}/${maxAttempts})`);
      }
      setTimeout(check, 50);
    }
  }
  
  check();
}

// æ™ºèƒ½åˆå§‹åŒ–
function initializeTest() {
  console.log("ğŸ” Checking environment and dependencies...");
  
  const isJSAR = isJSARRuntime();
  const hasAsyncLoader = typeof waitForWebGLTestGlobals !== 'undefined';
  const requiredGlobals = ['description', 'debug', 'WebGLTestUtils'];
  
  console.log("Environment info:", {
    isJSARRuntime: isJSAR,
    hasAsyncLoader: hasAsyncLoader,
    userAgent: typeof navigator !== 'undefined' ? navigator.userAgent : 'N/A'
  });
  
  if (isJSAR || hasAsyncLoader) {
    console.log("ğŸ”„ JSAR Runtime environment detected, using async loader...");
    
    if (hasAsyncLoader) {
      waitForWebGLTestGlobals(runWebGLTest, {
        maxRetries: 300,
        retryInterval: 50,
        timeout: 15000,
        onTimeout: function() {
          console.error("âŒ Timeout: WebGL test dependencies not available after 15s");
          try {
            if (typeof debug !== 'undefined') {
              debug("Test failed: Timeout waiting for dependencies");
            }
          } catch (e) {
            console.error("Debug function not available:", e);
          }
        },
        onError: function() {
          console.error("âŒ Error: Failed to load WebGL test dependencies");
          try {
            if (typeof debug !== 'undefined') {
              debug("Test failed: Failed to load dependencies");
            }
          } catch (e) {
            console.error("Debug function not available:", e);
          }
        }
      });
    } else {
      console.log("â³ No async loader found, using simple polling...");
      waitForGlobalsSimple(requiredGlobals, runWebGLTest);
    }
  } else {
    console.log("ğŸŒ Browser environment detected, running test directly...");
    setTimeout(runWebGLTest, 500);
  }
}

// ç¡®ä¿DOMåŠ è½½å®Œæˆåå†åˆå§‹åŒ–
setTimeout(initializeTest, 500);

</script>
<!-- <script src="../../js/js-test-post.js"></script> -->

</body>
</html>
